refactor(processor): reduce complexity and improve code organization
## Complexity Reduction

Reduce cognitive complexity in 5 functions to fix golangci-lint gocognit violations:

- **executeActionWithRecovery**: 40 → <20 (extract panic recovery, timeout handling helpers)
- **publishNewSpeciesDetectionEvent**: 27 → <20 (extract notification, event creation helpers)
- **New**: 23 → <20 (extract initialization helpers for deduplicator, tracker, client)
- **pendingDetectionsFlusher**: 22 → <20 (extract flush loop, config change logging)
- **persistDynamicThresholds**: 21 → <20 (extract conversion, cleanup, retry logic)

## Code Organization

Split 1414-line actions.go into 4 focused files (all under 500 lines):

- **actions_types.go** (285 lines): interfaces, types, and constants
- **actions_composite.go** (259 lines): CompositeAction implementation
- **actions_database.go** (422 lines): database and logging actions
- **actions_integrations.go** (476 lines): external service integrations

## Code Quality Improvements

- Extract DefaultFlushInterval constant (replaces magic number)
- Remove unnecessary mutex from SaveAudioAction (instances are local)
- Remove duplicate console output from LogAction (logger handles it)
- Add graceful shutdown to pendingDetectionsFlusher (context cancellation)
- Remove redundant error logging from SaveAudioAction (wrapper logs it)
- Fix misleading "Detection logged" message (only show on success)
- Fix context propagation in executeActionWithRecovery (chain from parent)
- Fix ScientificName persistence in dynamic thresholds (was missing)
- Document intentional error handling patterns
- Document legacy action goroutine leak limitation

No functional changes to detection pipeline behavior.
